+++ .env
@@ -1,0 +1,5 @@
+DB_HOST=172.18.0.2
+DB_PORT=5432
+POSTGRES_DB=devdb
+POSTGRES_USER=devuser
+POSTGRES_PASS=dimadom
