+++ app/rates/service.py
@@ -8,41 +8,6 @@ from app.database import Session
 from app.rates.utils import PortColumn, AggregateFunctions
 
 
-def fuzzy_search_port(
-        string_search: str, list_of_ports: list[dict], number_of_best_candidates: int
-) -> list[dict]:
-    """
-    Find the port that is most similar to the string_search by comparing code, name, parent slug using
-    fuzzy matching using the library rapidfuzz.
-
-        Parameters:
-                    string_search (str): The word you want to match with the ports
-                    list_of_ports (list[dict]): A list of port ports represented as a dictionary
-                    number_of_best_candidates (int): Limit the number of results
-
-        Returns:
-                best_results (list[List[dict]]): The n number
-
-    """
-
-    # The fuzzy search is applied for each of the dictionary of the list, and it will return a list of
-    # tuple for each key word of the dictionary. The tuples have the following structure
-    # (value of the key, score from 0 to 100, key of the value)
-    # the similarity fuzzy search is done by each value of the dictionary
-    list_of_scores = (
-        process.extract(string_search, dict(row), scorer=fuzz.WRatio)
-        for row in list_of_ports
-    )
-
-    # Then sort the tokens based on each of the value have the most score
-    # by taking the highest score of the list of tuples as the sort argument
-    sorted_scores = sorted(
-        list_of_scores,
-        key=lambda list_of_tuples: max(element[1] for element in list_of_tuples),
-        reverse=True,
-    )
-
-    return sorted_scores[:number_of_best_candidates]
 
 
 def generate_date_filter(
@@ -84,17 +49,11 @@ def generate_aggregate_funct(
 
 
 def generate_port_filter(column_type: PortColumn, column_name: str, value: str) -> str:
-    if column_type == PortColumn.NAME:
-        return f"and {column_name}.name = '{value}'"
-    if column_type == PortColumn.CODE:
-        return f"and {column_name}.code = '{value}'"
 
-    return ""
+    if column_type == PortColumn.NONE:
+        return ''
 
-
-def formatter(sequence: Iterable, element_modifier: Callable) -> Iterable:
-    return (element_modifier(element) for element in sequence)
-
+    return f"and {column_name}.{column_type.name} = '{value}'"
 
 def generate_rates_query(
         origin: Tuple[str, PortColumn] | None,
